refactor(cli): split flat multi-subcommand files into folder descent
alerts, notifiers, cloud-account, dlq, drift, and worker each had every
subcommand crammed into one file instead of following the repo's own
convention (<name>.rs dispatcher + <name>/<subcommand>.rs, same as app,
config, data, deploy, infra, release, sync, delete). Pure structural
split, no logic changes — each subcommand's tests moved with it.
