+++ cli/src/dlq.rs
@@ -2,7 +2,10 @@ use std::io::Write;
 
 use anyhow::{Context, Result, bail};
 use clap::{Arg, ArgAction, ArgMatches, Command};
+use remove::RemoveCommand;
+use retry::RetryCommand;
 use serde::{Deserialize, Serialize};
+use stats::StatsCommand;
 use termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
 
 use crate::{
@@ -10,6 +13,10 @@ use crate::{
     constants::{ERROR_FAILED_TO_SEND_REQUEST, get_platform_management_api_url},
     core::{command::command, http_client, validate::require_auth},
 };
+
+mod remove;
+mod retry;
+mod stats;
 
 // ── Top-level command ─────────────────────────────────────────────────────────
 
@@ -63,97 +70,6 @@ impl CliCommand for DlqCommand {
     }
 }
 
-// ── Retry sub-subcommand ──────────────────────────────────────────────────────
-
-#[derive(Debug)]
-struct RetryCommand;
-
-impl RetryCommand {
-    fn new() -> Self {
-        Self
-    }
-}
-
-impl CliCommand for RetryCommand {
-    fn command(&self) -> Command {
-        command("retry", "Retry a job from the dead-letter queue").arg(
-            Arg::new("job_id")
-                .required(true)
-                .help("The dead-letter job ID to retry"),
-        )
-    }
-
-    fn handler(&self, matches: &ArgMatches) -> Result<()> {
-        let _token = require_auth()?;
-        let job_id = matches
-            .get_one::<String>("job_id")
-            .context("job id is required")?;
-
-        let url = format!(
-            "{}/dlq/{}/retry",
-            get_platform_management_api_url(),
-            urlencoding::encode(job_id)
-        );
-        let response =
-            http_client::post(&url, serde_json::json!({})).with_context(|| ERROR_FAILED_TO_SEND_REQUEST)?;
-
-        if response.status().as_u16() == 404 {
-            bail!("DLQ job '{}' not found.", job_id);
-        }
-        if !response.status().is_success() {
-            bail!(
-                "Failed to retry DLQ job: {}",
-                response.text().unwrap_or_default()
-            );
-        }
-
-        print_action_result("Retrying", job_id)
-    }
-}
-
-// ── Remove sub-subcommand ─────────────────────────────────────────────────────
-
-#[derive(Debug)]
-struct RemoveCommand;
-
-impl RemoveCommand {
-    fn new() -> Self {
-        Self
-    }
-}
-
-impl CliCommand for RemoveCommand {
-    fn command(&self) -> Command {
-        command("remove", "Permanently remove a job from the dead-letter queue").arg(
-            Arg::new("job_id")
-                .required(true)
-                .help("The dead-letter job ID to remove"),
-        )
-    }
-
-    fn handler(&self, matches: &ArgMatches) -> Result<()> {
-        let _token = require_auth()?;
-        let job_id = matches
-            .get_one::<String>("job_id")
-            .context("job id is required")?;
-
-        let url = format!("{}/dlq/{}", get_platform_management_api_url(), urlencoding::encode(job_id));
-        let response = http_client::delete(&url).with_context(|| ERROR_FAILED_TO_SEND_REQUEST)?;
-
-        if response.status().as_u16() == 404 {
-            bail!("DLQ job '{}' not found.", job_id);
-        }
-        if !response.status().is_success() {
-            bail!(
-                "Failed to remove DLQ job: {}",
-                response.text().unwrap_or_default()
-            );
-        }
-
-        print_action_result("Removed", job_id)
-    }
-}
-
 fn print_action_result(verb: &str, job_id: &str) -> Result<()> {
     let mut stdout = StandardStream::stdout(ColorChoice::Always);
     stdout.set_color(ColorSpec::new().set_fg(Some(Color::Green)).set_bold(true))?;
@@ -161,53 +77,6 @@ fn print_action_result(verb: &str, job_id: &str) -> Result<()> {
     stdout.reset()?;
     writeln!(stdout, "  DLQ job {}", job_id)?;
     Ok(())
-}
-
-// ── Stats sub-subcommand ──────────────────────────────────────────────────────
-
-#[derive(Debug)]
-struct StatsCommand;
-
-impl StatsCommand {
-    fn new() -> Self {
-        Self
-    }
-}
-
-impl CliCommand for StatsCommand {
-    fn command(&self) -> Command {
-        command("stats", "Show dead-letter queue statistics")
-    }
-
-    fn handler(&self, matches: &ArgMatches) -> Result<()> {
-        let _token = require_auth()?;
-        let json_output = matches.get_flag("json");
-
-        let url = format!("{}/dlq/stats", get_platform_management_api_url());
-        let response = http_client::get(&url).with_context(|| ERROR_FAILED_TO_SEND_REQUEST)?;
-        if !response.status().is_success() {
-            bail!(
-                "Failed to get DLQ stats: {}",
-                response.text().unwrap_or_default()
-            );
-        }
-        let stats: DlqStats = response
-            .json()
-            .with_context(|| "Failed to parse DLQ stats response")?;
-
-        if json_output {
-            println!("{}", serde_json::to_string_pretty(&stats)?);
-            return Ok(());
-        }
-
-        let mut stdout = StandardStream::stdout(ColorChoice::Always);
-        writeln!(stdout)?;
-        writeln!(stdout, "  total       {}", stats.total)?;
-        writeln!(stdout, "  waiting     {}", stats.waiting)?;
-        writeln!(stdout, "  processed   {}", stats.processed)?;
-        writeln!(stdout)?;
-        Ok(())
-    }
 }
 
 // ── List handler ──────────────────────────────────────────────────────────────
@@ -302,13 +171,6 @@ struct DlqJob {
     failed_at: String,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
-struct DlqStats {
-    total: u32,
-    waiting: u32,
-    processed: u32,
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -325,26 +187,6 @@ mod tests {
     #[test]
     fn list_has_no_required_args() {
         assert!(dlq_cmd().try_get_matches_from(["dlq"]).is_ok());
-    }
-
-    #[test]
-    fn retry_requires_job_id() {
-        assert!(dlq_cmd().try_get_matches_from(["dlq", "retry"]).is_err());
-        assert!(
-            dlq_cmd()
-                .try_get_matches_from(["dlq", "retry", "job-1"])
-                .is_ok()
-        );
-    }
-
-    #[test]
-    fn remove_requires_job_id() {
-        assert!(dlq_cmd().try_get_matches_from(["dlq", "remove"]).is_err());
-        assert!(
-            dlq_cmd()
-                .try_get_matches_from(["dlq", "remove", "job-1"])
-                .is_ok()
-        );
     }
 
     #[test]
@@ -367,12 +209,5 @@ mod tests {
         let json = r#"[{"id":"job-1","deploymentId":"d","applicationId":"a","error":"e","attempts":1,"failedAt":"2024-01-15T10:00:00Z"}]"#;
         let jobs: Vec<DlqJob> = serde_json::from_str(json).unwrap();
         assert_eq!(jobs.len(), 1);
-    }
-
-    #[test]
-    fn dlq_stats_deserializes() {
-        let json = r#"{"total": 5, "waiting": 2, "processed": 3}"#;
-        let stats: DlqStats = serde_json::from_str(json).unwrap();
-        assert_eq!(stats.total, 5);
     }
 }
